requirements.txt
Flask==2.3.4
Flask-Cors==3.1.1
ultralytics==8.1.18
torch==2.1.0
torchvision==0.16.1
opencv-python==4.8.1.78
twilio==8.14.1
Pillow==10.1.0
